+++ src/ScottPlot5/ScottPlot5 Sandbox/Sandbox.WinForms/Form1.cs
@@ -1,4 +1,5 @@
 ﻿using ScottPlot;
+using ScottPlot.Control;
 
 namespace Sandbox.WinForms;
 
@@ -8,9 +9,14 @@ public partial class Form1 : Form
     {
         InitializeComponent();
 
-        formsPlot1.Plot.Add.Signal(Generate.Sin());
-        formsPlot1.Plot.Add.Signal(Generate.Cos());
-        formsPlot1.Plot.Axes.Rules.Add(new ScottPlot.AxisRules.SnapToTicksX(formsPlot1.Plot.Axes.Bottom));
-        formsPlot1.Plot.Axes.Rules.Add(new ScottPlot.AxisRules.SnapToTicksY(formsPlot1.Plot.Axes.Left));
+        var sig1 = formsPlot1.Plot.Add.Signal(Generate.Sin());
+        sig1.Axes.YAxis = formsPlot1.Plot.Axes.Left;
+        formsPlot1.Plot.Axes.Color(sig1.Axes.YAxis, sig1.Color);
+
+        var sig2 = formsPlot1.Plot.Add.Signal(Generate.Cos(mult: 1000));
+        sig2.Axes.YAxis = formsPlot1.Plot.Axes.Right;
+        formsPlot1.Plot.Axes.Color(sig2.Axes.YAxis, sig2.Color);
+
+        checkBox1.CheckedChanged += (s, e) => formsPlot1.Interaction.ChangeOpposingAxesTogether = checkBox1.Checked;
     }
 }