+++ src/ScottPlot5/ScottPlot5/Control/StandardZoomRectangle.cs
@@ -1,9 +1,10 @@
-﻿namespace ScottPlot.Control;
+﻿
+namespace ScottPlot.Control;
 
 /// <summary>
 /// Logic for drawing the shaded region on the plot when the user middle-click-drags to zoom
 /// </summary>
-public class StandardZoomRectangle : IZoomRectangle
+public class StandardZoomRectangle(Plot plot) : IZoomRectangle
 {
     public bool IsVisible { get; set; } = false;
 
@@ -15,36 +16,14 @@ public class StandardZoomRectangle : IZoomRectangle
     public Pixel MouseUp { get; set; }
     public bool HorizontalSpan { get; set; } = false;
     public bool VerticalSpan { get; set; } = false;
+    public Plot Plot { get; } = plot;
 
-    public void Apply(Plot plot)
-    {
-        PixelRect dataRect = plot.RenderManager.LastRender.DataRect;
-
-        IAxis? axisUnderMouse = plot.GetAxis(MouseDown);
-        if (axisUnderMouse is not null)
-        {
-            if (axisUnderMouse is IXAxis xAxis)
-            {
-                Apply(xAxis, dataRect);
-            }
-            else if (axisUnderMouse is IYAxis yAxis)
-            {
-                Apply(yAxis, dataRect);
-            }
-        }
-        else
-        {
-            plot.Axes.XAxes.ForEach(ax => Apply(ax, dataRect));
-            plot.Axes.YAxes.ForEach(ax => Apply(ax, dataRect));
-        }
-    }
-
-    private void Apply(IXAxis xAxis, PixelRect dataRect)
+    public void Apply(IXAxis xAxis)
     {
         if (HorizontalSpan == true)
-        {
             return;
-        }
+
+        PixelRect dataRect = Plot.RenderManager.LastRender.DataRect;
         double x1 = xAxis.GetCoordinate(MouseDown.X, dataRect);
         double x2 = xAxis.GetCoordinate(MouseUp.X, dataRect);
         double xMin = Math.Min(x1, x2);
@@ -52,12 +31,12 @@ public class StandardZoomRectangle : IZoomRectangle
         xAxis.Range.Set(xMin, xMax);
     }
 
-    private void Apply(IYAxis yAxis, PixelRect dataRect)
+    public void Apply(IYAxis yAxis)
     {
         if (VerticalSpan == true)
-        {
             return;
-        }
+
+        PixelRect dataRect = Plot.RenderManager.LastRender.DataRect;
         double y1 = yAxis.GetCoordinate(MouseDown.Y, dataRect);
         double y2 = yAxis.GetCoordinate(MouseUp.Y, dataRect);
         double xMin = Math.Min(y1, y2);